Use w3m to parse HTML code

+++ autoSUID.sh
@@ -3,7 +3,7 @@
 # Author: alb3rtov
 #
 # Descrition:
-# This script allows you to find SUID binaries and check if one of them 
+# This script allows you to find and enumerate SUID binaries and check if one of them 
 # can be used to escalate or mantain elevated privileges in a iteractive way.
 #
 
@@ -90,15 +90,33 @@ function check_dependencies() {
   sleep 2	
 	BLA::stop_loading_animation
 
-  if test -f $(which html2text); then
-		echo -e "${YELLOW}\n[*] Html2text is installed on the system (${NC}${LIGHTGREEN}V${NC}${YELLOW})${NC}"	
+  if test -f $(which curl); then
+    echo -e "${YELLOW}\n[*] Curl is installed on the system (${NC}${LIGHTGREEN}V${NC}${YELLOW})${NC}"	
 		sleep 2
-    return 0
+  else
+    echo -e "${YELLOW}\n[*] Curl is NOT installed on the system, exiting... (${NC}${LIGHTRED}X${NC}${YELLOW})${NC}"	
+		sleep 2
+    return 255
+  fi
+  
+  if test -f $(which html2text); then
+		echo -e "${YELLOW}\n[Optional] Html2text is installed on the system (${NC}${LIGHTGREEN}V${NC}${YELLOW})${NC}"	
+		sleep 2
+    return 0 
 	else
-		echo -e "${YELLOW}\n[*] Html2text is NOT installed on the system.\n\n [*] Some of the characters may not be displayed correctly, its recommended to install html2text ${NC}" 
+		echo -e "${YELLOW}\n[Optional] Html2text is NOT installed on the system.\n\n [*] Some of the characters may not be displayed correctly, its recommended to install html2text ${NC}" 
     sleep 2
-    return 1
 	fi
+  
+  if test -f $(which w3m); then
+    echo -e "${YELLOW}\n[Optional] W3m is installed on the system (${NC}${LIGHTGREEN}V${NC}${YELLOW})${NC}"	
+		sleep 2
+    return 1
+  else
+    echo -e "${YELLOW}\n[Optional] W3m is NOT installed on the system (${NC}${LIGHTRED}X${NC}${YELLOW})${NC}"	
+		sleep 2
+    return 2
+  fi
 }
 
 # Search and compare GTFO binaries with current SUID binaries
@@ -229,6 +247,9 @@ function suid_binaries_menu() {
 # Request to GTFObins info about selected SUID binary
 function request_bin_info() {
 	clear
+
+  local arg1=$1
+
 	if [ $selected_menu -eq 1 ]; then
 		selected_bin=${exploitable_binaries[${bin_index}]}
 		url_selected_bin=${url_exploitable_binaries[${bin_index}]}
@@ -237,7 +258,7 @@ function request_bin_info() {
 		url_selected_bin=${limited_url_exploitable_binaries[${bin_index}]}
 	fi
 
-	url="$1$url_selected_bin"
+	url="$arg1$url_selected_bin"
 
   echo -e "${YELLOW}\n[*] Searching info about $selected_bin... ${LIGHTPURPLE} ($url) ${NC}"
 	
@@ -249,10 +270,26 @@ function request_bin_info() {
 	fi
 }
 
+# Fix command suggestion interpreting HTML code with w3m
+function fixed_command() {
+  local arg1=$1
+  
+  cmd=""
+  num_lines=$(echo "$arg1" | wc -l)
+
+  for i in $(seq 1 $num_lines)
+  do
+    cmd+=$(echo "$arg1" | sed -n $(echo $i)p | w3m -dump -T text/html)
+    cmd+=$(echo "\n")
+  done
+}
+
 # Get description and command for SUID binary exploitation
 function extract_html_info() {
 
-  if [ $1 -eq 0 ]; then
+  local mode=$1
+
+  if [ $mode -eq 0 ]; then # Html2text
     if [ $selected_menu -eq 1 ]; then
 		  description=$(html2text output.html | grep -F "***** SUID *****" -A 50 | sed -n '/^\*\*\*\*\* SUID \*\*\*\*\*$/,/^\*\*\*\*\* Sudo \*\*\*\*\*$/p' | sed '1d;$d' | grep "*" -B 50 | sed '/*/d')
 		  commands=$(html2text output.html | grep -F "***** SUID *****" -A 50 | sed -n '/^\*\*\*\*\* SUID \*\*\*\*\*$/,/^\*\*\*\*\* Sudo \*\*\*\*\*$/p' | sed '1d;$d' | grep "*" -A 50 | sed 's/*//')
@@ -260,7 +297,19 @@ function extract_html_info() {
 		  description=$(html2text output.html | grep -F "***** Limited SUID *****" -A 50 | sed '1d;$d' | grep "*" -B 50 | sed '/*/d')
 		  commands=$(html2text output.html | grep -F "***** Limited SUID *****" -A 50 | sed '1d' | grep "*" -A 50 | sed 's/*//')
     fi
-  else
+  elif [ $mode -eq 1 ]; then # W3m
+    if [ $selected_menu -eq 1 ]; then
+      description=$(cat output.html | grep -A 25 'id="suid"' | awk '/<p>/,/<\/p>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
+      arg=$(cat output.html | grep -A 25 'id="suid"' | awk '/<pre>/,/<\/pre>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
+      fixed_command "$arg"
+      commands=$(echo $cmd)
+    else
+      description=$(cat output.html | grep -A 25 'id="limited-suid"' | awk '/<p>/,/<\/p>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
+      commands=$(cat output.html | grep -A 25 'id="limited-suid"' | awk '/<pre>/,/<\/pre>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
+      fixed_command "$arg"
+      commands=$(echo $cmd)
+    fi
+  else # Only curl
     if [ $selected_menu -eq 1 ]; then
       description=$(cat output.html | grep -A 25 'id="suid"' | awk '/<p>/,/<\/p>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
       commands=$(cat output.html | grep -A 25 'id="suid"' | awk '/<pre>/,/<\/pre>/' | sed -e 's/<[^>]*>//g' | sed -e 's/[ \t]*//')
@@ -286,13 +335,17 @@ function main() {
     gtfo_url="https://gtfobins.github.io"
     suid_urls=$(curl -s $gtfo_url -X GET | grep "#suid" | sed 's/<li><a href="//' | sed 's/">SUID<\/a><\/li>//')
     limited_suid_urls=$(curl -s $gtfo_url -X GET | grep "#limited-suid" | sed 's/<li><a href="//' | sed 's/">Limited SUID<\/a><\/li>//')
+    
     check_dependencies
     mode=$?
-    search_binaries
-	  display_menu
-	  request_bin_info $gtfo_url
-	  extract_html_info	$mode
-	  rm output.html 2> /dev/null
+    
+    if [ $mode -ne 255 ]; then
+      search_binaries
+	    display_menu
+	    request_bin_info $gtfo_url
+	    extract_html_info	$mode
+	    rm output.html 2> /dev/null
+    fi
   else
     echo -e "${YELLOW}\n[*] No internet connection, exiting...\n${NC}"
     tput cnorm
